cloudera-fraud-detection/requirements.txt
pyyaml>=6.0
python-dotenv>=0.20.0
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=1.0.0
matplotlib>=3.5.0 
flask>=2.0.0
lightgbm>=3.3.0
joblib>=1.2.0